+++ pkg/repository/scheduler_queue.go
@@ -639,6 +639,63 @@ func (d *queueRepository) GetDesiredLabels(ctx context.Context, tx *OptimisticTx
 	return stepIdToLabels, nil
 }
 
+func (d *queueRepository) GetStepSlotRequests(ctx context.Context, tx *OptimisticTx, stepIds []uuid.UUID) (map[uuid.UUID]map[string]int32, error) {
+	ctx, span := telemetry.NewSpan(ctx, "get-step-slot-requests")
+	defer span.End()
+
+	uniqueStepIds := sqlchelpers.UniqueSet(stepIds)
+
+	stepIdsToLookup := make([]uuid.UUID, 0, len(uniqueStepIds))
+	stepIdToRequests := make(map[uuid.UUID]map[string]int32, len(uniqueStepIds))
+
+	for _, stepId := range uniqueStepIds {
+		if value, found := d.stepIdSlotRequestsCache.Get(stepId); found {
+			stepIdToRequests[stepId] = value
+		} else {
+			stepIdsToLookup = append(stepIdsToLookup, stepId)
+		}
+	}
+
+	if len(stepIdsToLookup) == 0 {
+		return stepIdToRequests, nil
+	}
+
+	var queryTx sqlcv1.DBTX
+
+	if tx != nil {
+		queryTx = tx.tx
+	} else {
+		queryTx = d.pool
+	}
+
+	rows, err := d.queries.GetStepSlotRequests(ctx, queryTx, sqlcv1.GetStepSlotRequestsParams{
+		Stepids:  stepIdsToLookup,
+		Tenantid: d.tenantId,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	for _, row := range rows {
+		if _, ok := stepIdToRequests[row.StepID]; !ok {
+			stepIdToRequests[row.StepID] = make(map[string]int32)
+		}
+
+		stepIdToRequests[row.StepID][row.SlotType] = row.Units
+	}
+
+	// cache empty results so we skip DB lookups for steps without explicit slot requests
+	for _, stepId := range stepIdsToLookup {
+		if _, ok := stepIdToRequests[stepId]; !ok {
+			stepIdToRequests[stepId] = map[string]int32{}
+		}
+
+		d.stepIdSlotRequestsCache.Add(stepId, stepIdToRequests[stepId])
+	}
+
+	return stepIdToRequests, nil
+}
+
 func (d *queueRepository) RequeueRateLimitedItems(ctx context.Context, tenantId uuid.UUID, queueName string) ([]*sqlcv1.RequeueRateLimitedQueueItemsRow, error) {
 	tx, commit, rollback, err := sqlchelpers.PrepareTx(ctx, d.pool, d.l)