+++ sdks/python/hatchet_sdk/__init__.py
@@ -1,3 +1,4 @@
+from hatchet_sdk.cancellation import CancellationToken
 from hatchet_sdk.clients.admin import (
     RunStatus,
     ScheduleTriggerWorkflowOptions,
@@ -155,6 +156,8 @@ from hatchet_sdk.contracts.workflows_pb2 import (
     WorkerLabelComparator,
 )
 from hatchet_sdk.exceptions import (
+    CancellationReason,
+    CancelledError,
     DedupeViolationError,
     FailedTaskRunExceptionGroup,
     NonRetryableException,
@@ -177,6 +180,7 @@ from hatchet_sdk.runnables.workflow import TaskRunRef
 from hatchet_sdk.serde import is_in_hatchet_serialization_context
 from hatchet_sdk.utils.opentelemetry import OTelAttribute
 from hatchet_sdk.utils.serde import remove_null_unicode_character
+from hatchet_sdk.worker.slot_types import SlotType
 from hatchet_sdk.worker.worker import Worker, WorkerStartOptions, WorkerStatus
 from hatchet_sdk.workflow_run import WorkflowRunRef
 
@@ -193,6 +197,9 @@ __all__ = [
     "CELEvaluationResult",
     "CELFailure",
     "CELSuccess",
+    "CancellationReason",
+    "CancellationToken",
+    "CancelledError",
     "ClientConfig",
     "ClientTLSConfig",
     "ConcurrencyExpression",
@@ -254,6 +261,7 @@ __all__ = [
     "RunStatus",
     "ScheduleTriggerWorkflowOptions",
     "SleepCondition",
+    "SlotType",
     "StepRun",
     "StepRunDiff",
     "StepRunEventType",