+++ sdks/ruby/src/lib/hatchet/contracts/dispatcher/dispatcher_services_pb.rb
@@ -29,6 +29,10 @@ module Dispatcher
     rpc :RefreshTimeout, ::RefreshTimeoutRequest, ::RefreshTimeoutResponse
     rpc :ReleaseSlot, ::ReleaseSlotRequest, ::ReleaseSlotResponse
     rpc :UpsertWorkerLabels, ::UpsertWorkerLabelsRequest, ::UpsertWorkerLabelsResponse
+    # GetVersion returns the dispatcher protocol version as a simple integer.
+    # SDKs use this to determine feature support (e.g. slot_config registration).
+    # Old engines that do not implement this RPC will return UNIMPLEMENTED.
+    rpc :GetVersion, ::GetVersionRequest, ::GetVersionResponse
   end
 
   Stub = Service.rpc_stub_class